Refactoring - removed code from integer class and moved to numeric class
