+++ activesupport/lib/active_support/core_ext/numeric.rb
@@ -1,2 +1,3 @@
 require 'active_support/core_ext/numeric/bytes'
 require 'active_support/core_ext/numeric/time'
+require 'active_support/core_ext/numeric/multiple'
